Refactor scrolling text with logo sequence
